% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_module.R
\name{hist_ui}
\alias{hist_ui}
\title{Histogram Module UI}
\usage{
hist_ui(column_name, height = "300px")
}
\arguments{
\item{column_name}{Name of the column to create histogram for. Used as module ID.}

\item{height}{Height of the plot output. Default is "300px".}
}
\value{
A bslib card containing histogram output and range display
}
\description{
Creates a UI component for displaying a histogram with range information.
This module provides a reusable card component that shows a histogram plot
along with the data range in the header.
}
\examples{
\dontrun{
  # In your UI
  hist_ui("bill_length_mm")
  hist_ui("body_mass_g", height = "400px")
}
}
