% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Palmer Penguins Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny]{shinyApp}}}
}
\value{
Runs the Shiny application
}
\description{
Launch the Palmer Penguins Data Explorer Shiny application.
}
\examples{
\dontrun{
  run_app()
}
}
